+++ angular/src/app-initialize.ts
@@ -1,5 +1,5 @@
 import { NgZone } from '@angular/core';
-import { initialize } from '@ionic/core';
+import { setupConfig } from '@ionic/core';
 import { applyPolyfills, defineCustomElements } from '@ionic/core/loader';
 
 import { Config } from './providers/config';
@@ -11,7 +11,7 @@ export const appInitialize = (config: Config, doc: Document, zone: NgZone) => {
     const win: IonicWindow | undefined = doc.defaultView as any;
     if (win && typeof (window as any) !== 'undefined') {
 
-      initialize({
+      setupConfig({
         ...config,
         _zoneGate: (h: any) => zone.run(h)
       });