+++ packages/react-router/test-app/src/pages/routing/Details.tsx
@@ -8,9 +8,7 @@ import {
   IonButtons,
   IonBackButton,
   IonLabel,
-  IonCard,
   IonButton,
-  useIonViewWillEnter,
 } from '@ionic/react';
 import { useParams, useLocation } from 'react-router';